Update README.md
aslam o alaikum, I am a student of programming of html, css and javascript so if I you think that I am copying the project of someone 
so I am a student and I need to learn programming so I will not use copyied content in my work. but if you still think that i am copying so please 
let me use to learn programming and also I want to tell you that this project is incomplete and I will complete this after my school papers.